+++ src/lib/vl53lx/platform/src/vl53lx_platform_ipp.c
@@ -32,8 +32,8 @@ VL53LX_Error VL53LX_ipp_hist_process_data(
 	VL53LX_hist_post_process_config_t *ppost_cfg,
 	VL53LX_histogram_bin_data_t       *pbins,
 	VL53LX_xtalk_histogram_data_t     *pxtalk,
-	uint8_t                           *pArea1,
-	uint8_t                           *pArea2,
+	uint32_t                          *pArea1,
+	uint32_t                          *pArea2,
 	uint8_t                           *phisto_merge_nb,
 	VL53LX_range_results_t            *presults)
 {
